Implemented handle_wait to wait for events to happen on an object referred to by a handle (e.g. process death). Use handle_wait in startup/failshell to block until processes they run exit

+++ source/uspace/include/kernel/types.h
@@ -33,7 +33,7 @@ typedef signed long native_t;		/**< Signed native-sized type. */
 /** Kiwi-specific integer types. */
 typedef int32_t identifier_t;		/**< Type used to store an identifier for a global object. */
 typedef int32_t handle_t;		/**< Type used to store a handle to a per-process object. */
-typedef uint32_t timeout_t;		/**< Type used to store a timeout for an operation. */
+typedef int64_t timeout_t;		/**< Type used to store a timeout for an operation. */
 typedef int64_t offset_t;		/**< Type used to store an offset into something. */
 typedef uint64_t key_t;			/**< Type used to identify something. */
 typedef uint64_t file_size_t;		/**< Type used to store a file size. */
